Corte de etiqueta por faixa

+++ library/Wms/Service/Recebimento.php
@@ -64,7 +64,7 @@ class Recebimento
 
     public function retiraDigitoIdentificador($codigoBarras)
     {
-        $codigoBarras = (int) str_replace('.','',$codigoBarras);
+        $codigoBarras = (float) str_replace('.','',$codigoBarras);
         return substr($codigoBarras, 0, strlen($codigoBarras)-1);
     }
